+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * New cop `RescueModifier` tracks uses of `rescue` in modifier form
+* New cop `PercentLiterals` tracks uses of `%q`, `%Q`, `%s` and `%x`
 
 ### Bugs fixed